Split dependencies tests into several files
